npods helm charts execution fixes
- better affinity for non workloads
- apply helm charts later, and with a small artificial delay
  (fixes Kubernetes cluster unreachable: an error on the server
  ("unknown") has prevented the request from succeeding error)
- a bit more logic reshuffling to offest instabilities in initial
  scale operations
