+++ suites/npods/config/npods.yaml
@@ -1,3 +1,9 @@
+# NPods test suite configuration
+#
+# Configuration from this file is used only in the test suite for npods, where
+# values are used to configure how the tests run.
+#
+
 # Control cluster monitoring
 monitoring:
   # enable the loki helm chart
@@ -5,7 +11,7 @@ monitoring:
 
 workload:
   # default/template values for the npods helm workload yaml
-  default:
+  scale:
     name: 'npods-workload-default'
     replicas: 100
     image: 'msr.ci.mirantis.com/jnesbitt/n-pods-app:0.18'
@@ -14,7 +20,23 @@ workload:
     ram: '8192'
     threads: 100
 
-# parametrize that stability test
-stability_test:
-  duration: 600
-  period: 60
+  stability:
+    name: 'npods-workload-default'
+    replicas: 100
+    image: 'msr.ci.mirantis.com/jnesbitt/n-pods-app:0.18'
+    sleep: '10ms'
+    cpu: '300'
+    ram: '8192'
+    threads: 100
+
+
+tests:
+  # Control the scaling test
+  scale:
+    duration: 600
+    period: 60
+
+  # Configure the stability test
+  stability:
+    duration: 3600
+    period: 60
