+++ suites/npods/pytest.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-pytest -s --junitxml=./reports/junit.xml --html=./reports/report.html $@
+python -m pytest -s --junitxml=./reports/junit.xml --html=./reports/report.html $@
